Fixed errors and started analysis script
Fixed the following errors in the load_proj_data.R script: changed
REGEX to capture a variable name that was missing, fixed widths because
it missed a 4 char variable, and removed columns not relevant to the
study.  Also, loaded and prepared the data for inclusion in models by
figuring out which variables of the 340 are categorical.